cursor生成模拟数据和概览菜单完毕，改错完成

+++ core-service/src/main/java/com/ncg/dto/SimulationResponse.java
@@ -28,7 +28,7 @@ public class SimulationResponse {
     /**
      * 预警统计
      */
-    private Integer alertCount;
+    private Long alertCount;
 
     /**
      * 风险分布：{ Low: 数量, Medium: 数量, High: 数量 }
